Move lower section components down slightly
There is still a little bit more tightening that can be done vertically
on the PCB layout. Some of the components on the lower part of the board
have been moved down, to make space to move other components into.
The PCB width is acceptable, but reducing the height would be ideal.