Re-do layout, for v3.0
Having started the 3D CAD for the enclosure, it seems that the PCB is
able to be slightly wider than previously planned. To make the most of
this new width, a complete rearrange has been done - the PCB is now a
square shape.
Routing has been completed on this new layout, vias still need to be
added.